split stream_errors.go out of errors.go
